requirements.txt
python-dotenv==1.0.0
requests==2.30.0
terminaltables==3.1.10